+++ modules/main/src/main/java/com/nisovin/shopkeepers/ui/members/PlayerShopMembersEditorViewProvider.java
@@ -3,16 +3,12 @@ package com.nisovin.shopkeepers.ui.members;
 import org.bukkit.entity.Player;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopkeeper.player.members.DefaultPlayerShopAccessLevels;
-import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopkeeper;
 import com.nisovin.shopkeepers.ui.AbstractShopkeeperViewProvider;
 import com.nisovin.shopkeepers.ui.SKDefaultUITypes;
 import com.nisovin.shopkeepers.ui.lib.UIState;
 import com.nisovin.shopkeepers.ui.lib.View;
-import com.nisovin.shopkeepers.util.bukkit.PermissionUtils;
-import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 
 public class PlayerShopMembersEditorViewProvider extends AbstractShopkeeperViewProvider {
 
@@ -28,11 +24,9 @@ public class PlayerShopMembersEditorViewProvider extends AbstractShopkeeperViewP
 	@Override
 	public boolean canAccess(Player player, boolean silent) {
 		// Check access:
-		if (!this.getShopkeeper().hasAccessLevel(player, DefaultPlayerShopAccessLevels.FULL())
-				&& !PermissionUtils.hasPermission(player, ShopkeepersPlugin.BYPASS_PERMISSION)) {
+		if (!this.getShopkeeper().checkAccess(player, DefaultPlayerShopAccessLevels.FULL(), silent)) {
 			if (!silent) {
-				this.debugNotOpeningUI(player, "Missing member permission.");
-				TextUtils.sendMessage(player, Messages.notAllowedToEditShopMembers);
+				this.debugNotOpeningUI(player, "Player has no access.");
 			}
 			return false;
 		}